resmue-analyser-job-finder/requirements.txt
duckduckgo-search
phidata
groq
uvicorn 
fastapi
requests